refactor(core): extract runtime dump classes from types module
